+++ dbus-serialbattery/bms/jkbms_can.py
@@ -11,7 +11,7 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 from battery import Battery, Cell
-from utils import bytearray_to_string, get_connection_error_message, logger
+from utils import get_connection_error_message, generate_unique_identifier, logger
 from struct import unpack_from
 from time import sleep, time
 import sys
@@ -80,7 +80,7 @@ class Jkbms_Can(Battery):
         e.g. the serial number
         If there is no such value, please remove this function
         """
-        return self.port + ("__" + bytearray_to_string(self.address).replace("\\", "0") if self.address is not None else "")
+        return generate_unique_identifier(self.port, self.address)
 
     def test_connection(self):
         """