+++ dbus-serialbattery/bms/jkbms_pb.py
@@ -4,7 +4,7 @@
 # Added by https://github.com/KoljaWindeler
 
 from battery import Battery, Cell
-from utils import bytearray_to_string, read_serial_data, get_connection_error_message, logger, USE_PORT_AS_UNIQUE_ID
+from utils import read_serial_data, get_connection_error_message, logger
 from struct import unpack_from
 import sys
 
@@ -329,11 +329,7 @@ class Jkbms_pb(Battery):
         """
         Used to identify a BMS when multiple BMS are connected
         """
-        # TODO: Temporary solution, since the serial number is not correctly read
-        if USE_PORT_AS_UNIQUE_ID:
-            return self.port + ("__" + bytearray_to_string(self.address).replace("\\", "0") if self.address is not None else "")
-        else:
-            return self.unique_identifier_tmp
+        return self.unique_identifier_tmp
 
     def get_balancing(self):
         return 1 if self.balancing else 0